GuiInterface::getGuiInterface()->drawLine() now works properly

+++ gui/guiinterface.h
@@ -15,7 +15,7 @@ public:
     bool isOverride();
     std::vector<bool> isOverriddenBot();
     void show();
-    void drawPath(Point A, Point B, int seconds = 1);
+    void drawPath(Point A, Point B, double seconds = 1);
 
 private:
     static GuiInterface * gi;
